+++ MirrorGcode.py
@@ -29,23 +29,22 @@ class MirrorG():
         out_data = []
         mirror = False
         for line in src_data:
-            # - Find task block
             if not mirror:
                 # - Direct copy line
-                out_data.append(line + ("" if line.endswith("\r\n") or len(line) == 0 else "\r\n"))
+                out_data.append(line + ("" if line.endswith("\n") else "\n"))
  
-                # - Find task block
-                if re.match(r'^;.*(TASK BLOCK)', line):
+                # - Find first move command
+                if re.match('^(G0[01])', line):
                     mirror = True
                     continue
             else:
                 # - Replace rotation
-                rt = re.search('^(G0[01]) B([\-]{0,1}[0-9]+\.[0-9]+) F([0-9]+\.[0-9]+)', line)
+                rt = re.search(r'^(G0[01]) B([\-]{0,1}[0-9]+\.[0-9]+) F([0-9]+\.[0-9]+)', line)
                 if rt is not None:
                     CM = rt.group(1)
                     RT = float(rt.group(2))
                     FR = float(rt.group(3))
-                    out_data.append("%s B%.2f F%.1f\r\n" % (CM, -RT if RT != 0 else 0, FR))
+                    out_data.append("%s B%.2f F%.1f\n" % (CM, -RT if RT != 0 else 0, FR))
                     continue
 
                 withPowerChange = True
@@ -63,7 +62,7 @@ class MirrorG():
                         RY = float(mv.group(5))
                         FR = float(mv.group(6))
                         PW = float(mv.group(7))
-                        out_data.append("%s X%.2f Y%.2f Z%.2f A%.2f F%.1f S%.2f\r\n" % (CM, RX, RY, LX, LY, FR, PW))
+                        out_data.append("%s X%.2f Y%.2f Z%.2f A%.2f F%.1f S%.2f\n" % (CM, RX, RY, LX, LY, FR, PW))
                         continue
                     else:
                         CM = mv.group(1)
@@ -72,11 +71,11 @@ class MirrorG():
                         RX = float(mv.group(4))
                         RY = float(mv.group(5))
                         FR = float(mv.group(6))
-                        out_data.append("%s X%.2f Y%.2f Z%.2f A%.2f F%.1f\r\n" % (CM, RX, RY, LX, LY, FR))
+                        out_data.append("%s X%.2f Y%.2f Z%.2f A%.2f F%.1f\n" % (CM, RX, RY, LX, LY, FR))
                         continue
 
                 # - Direct copy line
-                out_data.append(line + ("" if line.endswith("\r\n") or len(line) == 0 else "\r\n"))
+                out_data.append(line + ("" if line.endswith("\n") else "\n"))
 
 
         fileName = file.replace(".gcode", "-mirror.gcode")